+++ Source/Preferences.h
@@ -12,9 +12,9 @@
 #define AUTO_PROPAGATE_TRANSLATION_SELECTED	1
 #define AUTO_PROPAGATE_TRANSLATION_ALL		2
 
-#define STARTUP_DO_NOTHING	0
+#define STARTUP_DO_NOTHING	    0
 #define STARTUP_OPEN_LAST_USED	1
-#define STARTUP_NEW_PROJET	2
+#define STARTUP_NEW_PROJET	    2
 #define STARTUP_PROJECT_BROWSER	3
 
 static NSString *const kAlwaysOverwriteGlossaryPrefs = @"always.overwrite.glossary";
@@ -22,22 +22,23 @@ static NSString *const kAlwaysViewGlossaryPrefs = @"always.view.glossary";
 
 @class StringEncoding;
 
-@interface Preferences : NSObject {
+@interface Preferences : NSObject
+{
 	NSMutableArray *mCachedIbtoolPlugins;
 }
 
 + (id)shared;
 
-- (int)maximumNumberOfHistoryFiles;
+- (NSInteger)maximumNumberOfHistoryFiles;
 - (BOOL)consoleDisplayUsingNSLog;
 
 - (BOOL)baseLanguageReadOnly;
 - (BOOL)developerMode;
 - (BOOL)ignoreCase;
 
-- (StringEncoding*)defaultEncoding;
+- (StringEncoding *)defaultEncoding;
 
 - (void)clearIbtoolPlugins;
-- (NSArray*)ibtoolPlugins;
+- (NSArray *)ibtoolPlugins;
 
 @end
